+++ CudaNeuralNetwork/LinearModel.cpp
@@ -4,20 +4,20 @@
 #include <iostream>
 #include <algorithm>
 
-LinearModel::LinearModel() 
+LinearModel::LinearModel()
 {
     std::srand(static_cast<unsigned int>(std::time(0)));
     m_weight = new double[3];
     std::generate(m_weight, m_weight + 3, []() { return ((static_cast<float>(rand()) / RAND_MAX) * 2.0f) - 1.0f; });
 }
 
-void LinearModel::training(float learning_rate, Vec2* training_data, int size, std::vector<double> point, std::vector<float>* error) 
+void LinearModel::training(double learning_rate, Vec2* training_data, int size, std::vector<double> point, std::vector<float>* error)
 {
     std::srand(static_cast<unsigned int>(std::time(0)));
     float errorCount = 0.0f;
-    int modulo = std::max(size,50);
+    int modulo = std::max(size, 50);
     float mf = modulo;
-    for (int i = 0; i < 1000000; i++) 
+    for (int i = 0; i < 1000000; i++)
     {
         int k = i % size;
         double yk = point[k];
@@ -30,9 +30,9 @@ void LinearModel::training(float learning_rate, Vec2* training_data, int size, s
         m_weight[2] += learning_rate * (yk - gXk) * xk.z;
         double p = predict(&training_data[k]);
         errorCount += (p >= 0 && yk >= 0) || (p < 0 && yk < 0) ? 0 : 100;
-        if (i % modulo == (modulo-1))
-        {            
-            error->push_back(errorCount/ mf);
+        if (i % modulo == (modulo - 1))
+        {
+            error->push_back(errorCount / mf);
             if ((errorCount / mf) <= 1.0f)
             {
                 break;
@@ -42,12 +42,12 @@ void LinearModel::training(float learning_rate, Vec2* training_data, int size, s
     }
 }
 
-double LinearModel::predict(Vec2* point) 
+double LinearModel::predict(Vec2* point)
 {
     return m_weight[0] + m_weight[1] * point->x + m_weight[2] * point->y;
 }
 
-LinearModel::~LinearModel() 
+LinearModel::~LinearModel()
 {
     delete[] m_weight;
 }
