+++ CudaNeuralNetwork/kernel.h
@@ -9,8 +9,8 @@
 #include "NeuralSwapData.hpp"
 
 void AddKernel(dim3 dimGrid, dim3 dimBlock,int* c, const int* a, const int* b, const int* size);
-void PropagateNeuralNetwork(dim3 dimGrid, dim3 dimBlock, const NeuralSwapData* nld, const NeuralNetworkDataCompact* nndc, const int* self_d, float*** self_w, float** self_x);
-void BackPropagateNeuralNetworkCompact(dim3 dimGrid, dim3 dimBlock, const NeuralSwapData* nld, const int* self_d, float*** self_w, float** self_x, float** self_delta);
-void BackPropagateNeuralNetworkCompactEnd(dim3 dimGrid, dim3 dimBlock, const NeuralSwapData* nld, const NeuralNetworkDataCompact* nndc, const int* self_d, float*** self_w, float** self_x, float** self_delta);
+void PropagateNeuralNetwork(dim3 dimGrid, dim3 dimBlock, const NeuralSwapData* nld, const NeuralNetworkDataCompact* nndc, const int* self_d, double*** self_w, double** self_x);
+void BackPropagateNeuralNetworkCompact(dim3 dimGrid, dim3 dimBlock, const NeuralSwapData* nld, const int* self_d, double*** self_w, double** self_x, double** self_delta);
+void BackPropagateNeuralNetworkCompactEnd(dim3 dimGrid, dim3 dimBlock, const NeuralSwapData* nld, const NeuralNetworkDataCompact* nndc, const int* self_d, double*** self_w, double** self_x, double** self_delta);
 
 #endif // KERNEL_H
